+++ clang/test/SemaCXX/lambda-capture-type-deduction.cpp
@@ -297,3 +297,26 @@ void __trans_tmp_1() {
 }
 
 }
+
+namespace GH47400 {
+
+struct Foo {};
+
+template <int, Foo f> struct Arr {};
+
+template <int> struct S {};
+
+constexpr bool foo() {
+  constexpr Foo f; // f constitutes an ODR-use
+  [&]<int is>() {
+    [&](Arr<is, f>) {}({});
+  }.template operator()<42>();
+
+  constexpr int C = 1;
+  []() {
+    [](S<C>) { }; // ... while C doesn't
+  };
+  return true;
+}
+
+} // namespace GH47400